Windows now playable!!
The port now runs end to end on Windows, verified by a full playthrough. Most of
what stood in the way was one bug class wearing different clothes: code that
reaches data by reproducing retail's link layout rather than by naming it. lld
orders and aligns globals differently from ld64, so these are latent on macOS and
fatal here.

Reach globals by name instead of by retail data layout. A mechanical sweep
resolved every hardcoded offset by looking the base symbol up in the decomp
object, adding the offset, and checking whether it landed on another symbol.
Seventeen sites resolved that way across tricky, texture, mm, shader, voxmaps,
gameloop, textrender, pi_dolphin, THPRead, BossDrakor, DLL 24, DLLs 7 and 60,
object 294 and DBstealerworm; offsets that landed inside their own symbol were
left alone. Three were doing real damage rather than feeding a no-op debug
print: THPRead was initialising three message queues at out-of-bounds addresses,
BossDrakor read its missile speed and lead factors out of an int reinterpreted as
float, and DLL 24 read particle scale factors two floats early. pi_videoinit
gains the 0x1B8-byte block the loading-screen strings live in, which the port had
never had at all, byte-for-byte from retail.

Two of these were the reported gameplay bugs. DBstealerworm addressed its
hit-reaction tables off an unrelated script table, so the damage byte for staff
priority 11 was struct padding and did nothing at all, while groundquake priority
17 read the 'O' of "popOutOfGround" as 79 damage and killed it outright; the
conflated gDbStealerwormSfxIds is now split into the three tables it always was,
with the damage table an explicit s8[] because reading a native int[] as bytes is
endian-dependent. SharpClaw read its family speed scale 5432 bytes past the end
of a 288-byte array, landing on pointers reinterpreted as floats, which fed
animPlaySpeed through root motion into velocity with only a lower clamp - they
teleported onto the player every frame instead of walking.

Free the object definition's own blocks in objFreeObjdef. gObjFileBufferTable
holds an ObjDef, but the entry was cast to GameObject, so the two frees took an
interior pointer into the ObjDef's own allocation and an address composed of half
of one pointer and half of another. Above 4GB both halves are non-zero, so the
null guards passed and a fabricated address went into the allocator on every
object-type unload. Retail frees ObjDef+0x30 and +0x34, which are modLines and
intersectionLines; all three frees are kept.

Stop poisoning a pointer through a retail offset. snowworm_applyReactionState
set INTERACT_FLAG_DISABLED by writing obj+0xAF, which is resetHitboxFlags in
retail but the top byte of anim.hitVolumeTransforms natively, so the camera's
targeting sweep dereferenced 0x0800... and crashed.

Pass a pointer where the callee wants one. KytesMum's update callback typedef
took an int while all three functions assigned to it take a GameObject, so the
truncated pointer faulted on the callee's first dereference.

Size objSetupLightChannels' light array from retail. The frame is 112 bytes with
the saved GPRs at r1+92 and the array at r1+52, so it holds ten pointers, not the
six declared; selectObjectLights clamps to the candidate count and never to the
caller's capacity.

Finish the Windows x64 argument work in the object DLLs, and bump aurora for a
NaN guard in the per-vertex lighting shader, where an unguarded 0/0 on
away-facing vertices rendered black under DXC but not under Metal.

+++ game/src/dlls/objects/614_KytesMum/KytesMum.c
@@ -309,7 +309,7 @@ void kytesmum_update(GameObject* obj)
     nearDist = 200.0f;
     if (runtime->questComplete == 0)
     {
-        if (runtime->updateCallback((int)obj) != 0)
+        if (runtime->updateCallback(obj) != 0)
         {
             mainSetBits(ObjAnim_ReadPlacementS16(&obj->anim, &(setup->completionGameBit)), 1);
             runtime->questComplete = 1;
